slice out refactor of augmenation realted stuff from new faction pr
* moved some of augmentation helper stuff to new file augmentation creator
* simplified init augmenations
* converted augmentation names to enum
