+++ src/Augmentation/data/AugmentationNames.ts
@@ -1,234 +1,121 @@
-export const AugmentationNames: {
-  Targeting1: string;
-  Targeting2: string;
-  Targeting3: string;
-  SyntheticHeart: string;
-  SynfibrilMuscle: string;
-  CombatRib1: string;
-  CombatRib2: string;
-  CombatRib3: string;
-  NanofiberWeave: string;
-  SubdermalArmor: string;
-  WiredReflexes: string;
-  GrapheneBoneLacings: string;
-  BionicSpine: string;
-  GrapheneBionicSpine: string;
-  BionicLegs: string;
-  GrapheneBionicLegs: string;
-  SpeechProcessor: string;
-  TITN41Injection: string;
-  EnhancedSocialInteractionImplant: string;
-  BitWire: string;
-  ArtificialBioNeuralNetwork: string;
-  ArtificialSynapticPotentiation: string;
-  EnhancedMyelinSheathing: string;
-  SynapticEnhancement: string;
-  NeuralRetentionEnhancement: string;
-  DataJack: string;
-  ENM: string;
-  ENMCore: string;
-  ENMCoreV2: string;
-  ENMCoreV3: string;
-  ENMAnalyzeEngine: string;
-  ENMDMA: string;
-  Neuralstimulator: string;
-  NeuralAccelerator: string;
-  CranialSignalProcessorsG1: string;
-  CranialSignalProcessorsG2: string;
-  CranialSignalProcessorsG3: string;
-  CranialSignalProcessorsG4: string;
-  CranialSignalProcessorsG5: string;
-  NeuronalDensification: string;
-  NeuroreceptorManager: string;
-  NuoptimalInjectorImplant: string;
-  SpeechEnhancement: string;
-  FocusWire: string;
-  PCDNI: string;
-  PCDNIOptimizer: string;
-  PCDNINeuralNetwork: string;
-  PCMatrix: string;
-  ADRPheromone1: string;
-  ADRPheromone2: string;
-  ShadowsSimulacrum: string;
-  HacknetNodeCPUUpload: string;
-  HacknetNodeCacheUpload: string;
-  HacknetNodeNICUpload: string;
-  HacknetNodeKernelDNI: string;
-  HacknetNodeCoreDNI: string;
-  NeuroFluxGovernor: string;
-  Neurotrainer1: string;
-  Neurotrainer2: string;
-  Neurotrainer3: string;
-  Hypersight: string;
-  LuminCloaking1: string;
-  LuminCloaking2: string;
-  HemoRecirculator: string;
-  SmartSonar: string;
-  PowerRecirculator: string;
-  QLink: string;
-  TheRedPill: string;
-  SPTN97: string;
-  HiveMind: string;
-  CordiARCReactor: string;
-  SmartJaw: string;
-  Neotra: string;
-  Xanipher: string;
-  nextSENS: string;
-  OmniTekInfoLoad: string;
-  PhotosyntheticCells: string;
-  Neurolink: string;
-  TheBlackHand: string;
-  UnstableCircadianModulator: string;
-  CRTX42AA: string;
-  Neuregen: string;
-  CashRoot: string;
-  NutriGen: string;
-  INFRARet: string;
-  DermaForce: string;
-  GrapheneBrachiBlades: string;
-  GrapheneBionicArms: string;
-  BrachiBlades: string;
-  BionicArms: string;
-  SNA: string;
-  HydroflameLeftArm: string;
-  EsperEyewear: string;
-  EMS4Recombination: string;
-  OrionShoulder: string;
-  HyperionV1: string;
-  HyperionV2: string;
-  GolemSerum: string;
-  VangelisVirus: string;
-  VangelisVirus3: string;
-  INTERLINKED: string;
-  BladeRunner: string;
-  BladeArmor: string;
-  BladeArmorPowerCells: string;
-  BladeArmorEnergyShielding: string;
-  BladeArmorUnibeam: string;
-  BladeArmorOmnibeam: string;
-  BladeArmorIPU: string;
-  BladesSimulacrum: string;
-  StaneksGift1: string;
-  StaneksGift2: string;
-  StaneksGift3: string;
-} = {
-  Targeting1: "Augmented Targeting I",
-  Targeting2: "Augmented Targeting II",
-  Targeting3: "Augmented Targeting III",
-  SyntheticHeart: "Synthetic Heart",
-  SynfibrilMuscle: "Synfibril Muscle",
-  CombatRib1: "Combat Rib I",
-  CombatRib2: "Combat Rib II",
-  CombatRib3: "Combat Rib III",
-  NanofiberWeave: "Nanofiber Weave",
-  SubdermalArmor: "NEMEAN Subdermal Weave",
-  WiredReflexes: "Wired Reflexes",
-  GrapheneBoneLacings: "Graphene Bone Lacings",
-  BionicSpine: "Bionic Spine",
-  GrapheneBionicSpine: "Graphene Bionic Spine Upgrade",
-  BionicLegs: "Bionic Legs",
-  GrapheneBionicLegs: "Graphene Bionic Legs Upgrade",
-  SpeechProcessor: "Speech Processor Implant",
-  TITN41Injection: "TITN-41 Gene-Modification Injection",
-  EnhancedSocialInteractionImplant: "Enhanced Social Interaction Implant",
-  BitWire: "BitWire",
-  ArtificialBioNeuralNetwork: "Artificial Bio-neural Network Implant",
-  ArtificialSynapticPotentiation: "Artificial Synaptic Potentiation",
-  EnhancedMyelinSheathing: "Enhanced Myelin Sheathing",
-  SynapticEnhancement: "Synaptic Enhancement Implant",
-  NeuralRetentionEnhancement: "Neural-Retention Enhancement",
-  DataJack: "DataJack",
-  ENM: "Embedded Netburner Module",
-  ENMCore: "Embedded Netburner Module Core Implant",
-  ENMCoreV2: "Embedded Netburner Module Core V2 Upgrade",
-  ENMCoreV3: "Embedded Netburner Module Core V3 Upgrade",
-  ENMAnalyzeEngine: "Embedded Netburner Module Analyze Engine",
-  ENMDMA: "Embedded Netburner Module Direct Memory Access Upgrade",
-  Neuralstimulator: "Neuralstimulator",
-  NeuralAccelerator: "Neural Accelerator",
-  CranialSignalProcessorsG1: "Cranial Signal Processors - Gen I",
-  CranialSignalProcessorsG2: "Cranial Signal Processors - Gen II",
-  CranialSignalProcessorsG3: "Cranial Signal Processors - Gen III",
-  CranialSignalProcessorsG4: "Cranial Signal Processors - Gen IV",
-  CranialSignalProcessorsG5: "Cranial Signal Processors - Gen V",
-  NeuronalDensification: "Neuronal Densification",
-  NeuroreceptorManager: "Neuroreceptor Management Implant",
-  NuoptimalInjectorImplant: "Nuoptimal Nootropic Injector Implant",
-  SpeechEnhancement: "Speech Enhancement",
-  FocusWire: "FocusWire",
-  PCDNI: "PC Direct-Neural Interface",
-  PCDNIOptimizer: "PC Direct-Neural Interface Optimization Submodule",
-  PCDNINeuralNetwork: "PC Direct-Neural Interface NeuroNet Injector",
-  PCMatrix: "PCMatrix",
-  ADRPheromone1: "ADR-V1 Pheromone Gene",
-  ADRPheromone2: "ADR-V2 Pheromone Gene",
-  ShadowsSimulacrum: "The Shadow's Simulacrum",
-  HacknetNodeCPUUpload: "Hacknet Node CPU Architecture Neural-Upload",
-  HacknetNodeCacheUpload: "Hacknet Node Cache Architecture Neural-Upload",
-  HacknetNodeNICUpload: "Hacknet Node NIC Architecture Neural-Upload",
-  HacknetNodeKernelDNI: "Hacknet Node Kernel Direct-Neural Interface",
-  HacknetNodeCoreDNI: "Hacknet Node Core Direct-Neural Interface",
-  NeuroFluxGovernor: "NeuroFlux Governor",
-  Neurotrainer1: "Neurotrainer I",
-  Neurotrainer2: "Neurotrainer II",
-  Neurotrainer3: "Neurotrainer III",
-  Hypersight: "HyperSight Corneal Implant",
-  LuminCloaking1: "LuminCloaking-V1 Skin Implant",
-  LuminCloaking2: "LuminCloaking-V2 Skin Implant",
-  HemoRecirculator: "HemoRecirculator",
-  SmartSonar: "SmartSonar Implant",
-  PowerRecirculator: "Power Recirculation Core",
-  QLink: "QLink",
-  TheRedPill: "The Red Pill",
-  SPTN97: "SPTN-97 Gene Modification",
-  HiveMind: "ECorp HVMind Implant",
-  CordiARCReactor: "CordiARC Fusion Reactor",
-  SmartJaw: "SmartJaw",
-  Neotra: "Neotra",
-  Xanipher: "Xanipher",
-  nextSENS: "nextSENS Gene Modification",
-  OmniTekInfoLoad: "OmniTek InfoLoad",
-  PhotosyntheticCells: "Photosynthetic Cells",
-  Neurolink: "BitRunners Neurolink",
-  TheBlackHand: "The Black Hand",
-  UnstableCircadianModulator: "Unstable Circadian Modulator",
-  CRTX42AA: "CRTX42-AA Gene Modification",
-  Neuregen: "Neuregen Gene Modification",
-  CashRoot: "CashRoot Starter Kit",
-  NutriGen: "NutriGen Implant",
-  INFRARet: "INFRARET Enhancement",
-  DermaForce: "DermaForce Particle Barrier",
-  GrapheneBrachiBlades: "Graphene BrachiBlades Upgrade",
-  GrapheneBionicArms: "Graphene Bionic Arms Upgrade",
-  BrachiBlades: "BrachiBlades",
-  BionicArms: "Bionic Arms",
-  SNA: "Social Negotiation Assistant (S.N.A)",
-  HydroflameLeftArm: "Hydroflame Left Arm",
-  EsperEyewear: "EsperTech Bladeburner Eyewear",
-  EMS4Recombination: "EMS-4 Recombination",
-  OrionShoulder: "ORION-MKIV Shoulder",
-  HyperionV1: "Hyperion Plasma Cannon V1",
-  HyperionV2: "Hyperion Plasma Cannon V2",
-  GolemSerum: "GOLEM Serum",
-  VangelisVirus: "Vangelis Virus",
-  VangelisVirus3: "Vangelis Virus 3.0",
-  INTERLINKED: "I.N.T.E.R.L.I.N.K.E.D",
-  BladeRunner: "Blade's Runners",
-  BladeArmor: "BLADE-51b Tesla Armor",
-  BladeArmorPowerCells: "BLADE-51b Tesla Armor: Power Cells Upgrade",
-  BladeArmorEnergyShielding: "BLADE-51b Tesla Armor: Energy Shielding Upgrade",
-  BladeArmorUnibeam: "BLADE-51b Tesla Armor: Unibeam Upgrade",
-  BladeArmorOmnibeam: "BLADE-51b Tesla Armor: Omnibeam Upgrade",
-  BladeArmorIPU: "BLADE-51b Tesla Armor: IPU Upgrade",
-  BladesSimulacrum: "The Blade's Simulacrum",
+export enum AugmentationNames {
+  Targeting1 = "Augmented Targeting I",
+  Targeting2 = "Augmented Targeting II",
+  Targeting3 = "Augmented Targeting III",
+  SyntheticHeart = "Synthetic Heart",
+  SynfibrilMuscle = "Synfibril Muscle",
+  CombatRib1 = "Combat Rib I",
+  CombatRib2 = "Combat Rib II",
+  CombatRib3 = "Combat Rib III",
+  NanofiberWeave = "Nanofiber Weave",
+  SubdermalArmor = "NEMEAN Subdermal Weave",
+  WiredReflexes = "Wired Reflexes",
+  GrapheneBoneLacings = "Graphene Bone Lacings",
+  BionicSpine = "Bionic Spine",
+  GrapheneBionicSpine = "Graphene Bionic Spine Upgrade",
+  BionicLegs = "Bionic Legs",
+  GrapheneBionicLegs = "Graphene Bionic Legs Upgrade",
+  SpeechProcessor = "Speech Processor Implant",
+  TITN41Injection = "TITN-41 Gene-Modification Injection",
+  EnhancedSocialInteractionImplant = "Enhanced Social Interaction Implant",
+  BitWire = "BitWire",
+  ArtificialBioNeuralNetwork = "Artificial Bio-neural Network Implant",
+  ArtificialSynapticPotentiation = "Artificial Synaptic Potentiation",
+  EnhancedMyelinSheathing = "Enhanced Myelin Sheathing",
+  SynapticEnhancement = "Synaptic Enhancement Implant",
+  NeuralRetentionEnhancement = "Neural-Retention Enhancement",
+  DataJack = "DataJack",
+  ENM = "Embedded Netburner Module",
+  ENMCore = "Embedded Netburner Module Core Implant",
+  ENMCoreV2 = "Embedded Netburner Module Core V2 Upgrade",
+  ENMCoreV3 = "Embedded Netburner Module Core V3 Upgrade",
+  ENMAnalyzeEngine = "Embedded Netburner Module Analyze Engine",
+  ENMDMA = "Embedded Netburner Module Direct Memory Access Upgrade",
+  Neuralstimulator = "Neuralstimulator",
+  NeuralAccelerator = "Neural Accelerator",
+  CranialSignalProcessorsG1 = "Cranial Signal Processors - Gen I",
+  CranialSignalProcessorsG2 = "Cranial Signal Processors - Gen II",
+  CranialSignalProcessorsG3 = "Cranial Signal Processors - Gen III",
+  CranialSignalProcessorsG4 = "Cranial Signal Processors - Gen IV",
+  CranialSignalProcessorsG5 = "Cranial Signal Processors - Gen V",
+  NeuronalDensification = "Neuronal Densification",
+  NeuroreceptorManager = "Neuroreceptor Management Implant",
+  NuoptimalInjectorImplant = "Nuoptimal Nootropic Injector Implant",
+  SpeechEnhancement = "Speech Enhancement",
+  FocusWire = "FocusWire",
+  PCDNI = "PC Direct-Neural Interface",
+  PCDNIOptimizer = "PC Direct-Neural Interface Optimization Submodule",
+  PCDNINeuralNetwork = "PC Direct-Neural Interface NeuroNet Injector",
+  PCMatrix = "PCMatrix",
+  ADRPheromone1 = "ADR-V1 Pheromone Gene",
+  ADRPheromone2 = "ADR-V2 Pheromone Gene",
+  ShadowsSimulacrum = "The Shadow's Simulacrum",
+  HacknetNodeCPUUpload = "Hacknet Node CPU Architecture Neural-Upload",
+  HacknetNodeCacheUpload = "Hacknet Node Cache Architecture Neural-Upload",
+  HacknetNodeNICUpload = "Hacknet Node NIC Architecture Neural-Upload",
+  HacknetNodeKernelDNI = "Hacknet Node Kernel Direct-Neural Interface",
+  HacknetNodeCoreDNI = "Hacknet Node Core Direct-Neural Interface",
+  NeuroFluxGovernor = "NeuroFlux Governor",
+  Neurotrainer1 = "Neurotrainer I",
+  Neurotrainer2 = "Neurotrainer II",
+  Neurotrainer3 = "Neurotrainer III",
+  Hypersight = "HyperSight Corneal Implant",
+  LuminCloaking1 = "LuminCloaking-V1 Skin Implant",
+  LuminCloaking2 = "LuminCloaking-V2 Skin Implant",
+  HemoRecirculator = "HemoRecirculator",
+  SmartSonar = "SmartSonar Implant",
+  PowerRecirculator = "Power Recirculation Core",
+  QLink = "QLink",
+  TheRedPill = "The Red Pill",
+  SPTN97 = "SPTN-97 Gene Modification",
+  HiveMind = "ECorp HVMind Implant",
+  CordiARCReactor = "CordiARC Fusion Reactor",
+  SmartJaw = "SmartJaw",
+  Neotra = "Neotra",
+  Xanipher = "Xanipher",
+  nextSENS = "nextSENS Gene Modification",
+  OmniTekInfoLoad = "OmniTek InfoLoad",
+  PhotosyntheticCells = "Photosynthetic Cells",
+  Neurolink = "BitRunners Neurolink",
+  TheBlackHand = "The Black Hand",
+  UnstableCircadianModulator = "Unstable Circadian Modulator",
+  CRTX42AA = "CRTX42-AA Gene Modification",
+  Neuregen = "Neuregen Gene Modification",
+  CashRoot = "CashRoot Starter Kit",
+  NutriGen = "NutriGen Implant",
+  INFRARet = "INFRARET Enhancement",
+  DermaForce = "DermaForce Particle Barrier",
+  GrapheneBrachiBlades = "Graphene BrachiBlades Upgrade",
+  GrapheneBionicArms = "Graphene Bionic Arms Upgrade",
+  BrachiBlades = "BrachiBlades",
+  BionicArms = "Bionic Arms",
+  SNA = "Social Negotiation Assistant (S.N.A)",
+  HydroflameLeftArm = "Hydroflame Left Arm",
+  EsperEyewear = "EsperTech Bladeburner Eyewear",
+  EMS4Recombination = "EMS-4 Recombination",
+  OrionShoulder = "ORION-MKIV Shoulder",
+  HyperionV1 = "Hyperion Plasma Cannon V1",
+  HyperionV2 = "Hyperion Plasma Cannon V2",
+  GolemSerum = "GOLEM Serum",
+  VangelisVirus = "Vangelis Virus",
+  VangelisVirus3 = "Vangelis Virus 3.0",
+  INTERLINKED = "I.N.T.E.R.L.I.N.K.E.D",
+  BladeRunner = "Blade's Runners",
+  BladeArmor = "BLADE-51b Tesla Armor",
+  BladeArmorPowerCells = "BLADE-51b Tesla Armor: Power Cells Upgrade",
+  BladeArmorEnergyShielding = "BLADE-51b Tesla Armor: Energy Shielding Upgrade",
+  BladeArmorUnibeam = "BLADE-51b Tesla Armor: Unibeam Upgrade",
+  BladeArmorOmnibeam = "BLADE-51b Tesla Armor: Omnibeam Upgrade",
+  BladeArmorIPU = "BLADE-51b Tesla Armor: IPU Upgrade",
+  BladesSimulacrum = "The Blade's Simulacrum",
 
-  StaneksGift1: "Stanek's Gift - Genesis",
-  StaneksGift2: "Stanek's Gift - Awakening",
-  StaneksGift3: "Stanek's Gift - Serenity",
+  StaneksGift1 = "Stanek's Gift - Genesis",
+  StaneksGift2 = "Stanek's Gift - Awakening",
+  StaneksGift3 = "Stanek's Gift - Serenity",
 
   //Wasteland Augs
   //PepBoy:                             "P.E.P-Boy", Plasma Energy Projection System
   //PepBoyForceField Generates plasma force fields
   //PepBoyBlasts Generate high density plasma concussive blasts
   //PepBoyDataStorage STore more data on pep boy,
-};
+}
