+++ src/Faction/ui/AugmentationsPage.tsx
@@ -23,7 +23,7 @@ import Typography from "@mui/material/Typography";
 import Tooltip from "@mui/material/Tooltip";
 import TableBody from "@mui/material/TableBody";
 import Table from "@mui/material/Table";
-import { CONSTANTS } from "../../Constants";
+import { getGenericAugmentationPriceMultiplier } from "../../Augmentation/AugmentationHelpers";
 
 type IProps = {
   faction: Faction;
@@ -180,10 +180,7 @@ export function AugmentationsPage(props: IProps): React.ReactElement {
       </>
     );
   }
-  const mult = Math.pow(
-    CONSTANTS.MultipleAugMultiplier * [1, 0.96, 0.94, 0.93][player.sourceFileLvl(11)],
-    player.queuedAugmentations.length,
-  );
+
   return (
     <>
       <Button onClick={props.routeToMainPage}>Back</Button>
@@ -204,7 +201,9 @@ export function AugmentationsPage(props: IProps): React.ReactElement {
             </Typography>
           }
         >
-          <Typography>Price multiplier: x {numeralWrapper.formatMultiplier(mult)}</Typography>
+          <Typography>
+            Price multiplier: x {numeralWrapper.formatMultiplier(getGenericAugmentationPriceMultiplier())}
+          </Typography>
         </Tooltip>
       </Box>
       <Button onClick={() => switchSortOrder(PurchaseAugmentationsOrderSetting.Cost)}>Sort by Cost</Button>
