+++ src/PersonObjects/Resleeving/Resleeving.ts
@@ -111,7 +111,7 @@ export function generateResleeves(): Resleeve[] {
         AugmentationNames.StaneksGift1,
         AugmentationNames.StaneksGift2,
         AugmentationNames.StaneksGift3,
-      ];
+      ].map((augmentation) => augmentation as string);
       if (forbidden.includes(randKey)) {
         continue;
       }
